# Backend/src/main/java/com/omerozturk/N11GraduationProject/csr/entities/dtos/CsrCustomerCreditSaveRequestDto.java
package com.omerozturk.N11GraduationProject.csr.entities.dtos;

import com.omerozturk.N11GraduationProject.csr.entities.enums.EnumCreditResult;
import com.omerozturk.N11GraduationProject.gen.utilities.enums.EnumStatus;
import lombok.Data;

import javax.validation.constraints.NotEmpty;
import java.math.BigDecimal;
import java.util.Date;

@Data
public class CsrCustomerCreditSaveRequestDto {
    private Long crdCreditId;
    private Long csrCustomerId;

}